[Untested] change the database read and write to float for SCurves
In accordance with the previous commit also change the data written
and read to/from the database for SCurves to also be floats.
NOTE: this a breaking change with the previous database. We won't be
able to read from the old database anymore, since it will crash when
trying to read an int dataset as a float.

We could introduce something similar to what is done in arraymancer's
HDF5 interface, where we make use of nimhdf5's `convertType` proc.

+++ InGridDatabase/src/ingridDatabase/databaseRead.nim
@@ -29,9 +29,9 @@ proc getScurve*[T: SomeInteger](h5f: var H5FileObj,
                           result.name)
   result.voltage = voltage
   var dset = h5f[dsetName.dset_str]
-  let data = dset[int64].reshape2D(dset.shape).transpose
+  let data = dset[float64].reshape2D(dset.shape).transpose
   result.thl = data[0].asType(int)
-  result.hits = data[1].asType(int)
+  result.hits = data[1]
 
 proc getScurve*(chipName: string, voltage: int): SCurve =
   ## reads the given voltages' SCurve for `chipName`
